fix: DeepSeek建议验证逻辑三合一修复
1. Flat评分双标修正: deepseek_advice_correct INT→FLOAT, 增加0.5部分正确
   - increase+flat=0.5(与advice_accuracy一致)
   - decrease+flat=0.5(与advice_accuracy一致)
   - hold+up=0.5(与advice_accuracy一致)

2. AI历史准确率回传prompt: 查询avg(deepseek_advice_correct)+count
   - prompt新增"AI建议准确率XX%(N次)"
   - 只给数据不给行为指令,避免自我实现预言

3. Gate覆盖时用AI原始方向回验: 解析ext_text1.raw_action
   - deepseek_advice_correct始终测AI真实判断
   - 即使Gate强制系统方向,仍评估AI原始方向是否正确
   - 能看清AI在极端行情前的真实捕捉能力

DDL: ALTER TABLE strategy_validation_log MODIFY COLUMN deepseek_advice_correct FLOAT

+++ backend/app/models/strategy_validation_log.py
@@ -207,7 +207,7 @@ class StrategyValidationLog(Base, TimestampMixin):
     deepseek_advice_action: Mapped[Optional[str]] = mapped_column(
         String(20), nullable=True, comment="AI建议方向(increase/hold/decrease)",
     )
-    deepseek_advice_correct: Mapped[Optional[int]] = mapped_column(
+    deepseek_advice_correct: Mapped[Optional[float]] = mapped_column(
         Integer, nullable=True, index=True, comment="AI建议准确度(0/1, T+1回填)",
     )
